==> ipynb/3_library_control_var_And_dependent_var.ipynb <==
import pandas as pd
import numpy as np

from glob import glob
from tqdm import tqdm

sido_abbr_nm = {
    '서울' : '서울특별시',
    '부산' : '부산광역시',
    '대구' : '대구광역시',
    '인천' : '인천광역시',
    '광주' : '광주광역시',
    '대전' : '대전광역시',
    '울산' : '울산광역시',
    '세종' : '세종특별자치시',
    '경기' : '경기도',
    '강원' : '강원도',
    '충북' : '충청북도',
    '충남' : '충청남도',
    '전북' : '전라북도',
    '전남' : '전라남도',
    '경북' : '경상북도',
    '경남' : '경상남도',
    '제주' : '제주특별자치도',
    '제주도' : '제주특별자치도'
}

# 시군구 지명 바꾸기
def city_sgg_change(sido,sgg):
    if (sido =='인천광역시')&(sgg=='남구'):
        sgg='미추홀구'
    elif (sido =='경기도')&(sgg=='여주군'):
        sgg ='여주시'
    elif (sido =='충청북도')&(sgg=='청원군'):
        sgg ='충주시'
    elif (sido =='세종특별자치시')&(sgg!=sgg):
        sgg ='세종특별자치시'
    elif (sido =='세종특별자치시')&(sgg=='세종시'):
        sgg ='세종특별자치시'
    elif (sido =='세종특별자치시')&(sgg=='연기군'):
        sgg ='세종특별자치시'
     
    return sgg.split(' ')[0] #경기도 구 단위를 시단위로 올리기 위해

## 대상 시군구

import geopandas as gpd

sgg_gdf = gpd.gpd.GeoDataFrame.from_file('../../data/0/shp/sgg.shp', encoding='UTF-8', crs="epsg:4326")
sgg_gdf = sgg_gdf.astype({'SGG_AD_CD':str,'SGG_LE_CD':str})

sgg_df = sgg_gdf[['SGG_AD_CD','SIDO','SGG']].copy()
sgg_df = sgg_df.rename(columns={'SGG_AD_CD':'시군구코드','SIDO':'시도','SGG':'시군구'})

sgg_df

## 도서관 데이터

lib_df_list =[]

for year in tqdm(range(2014,2019)):
    lib_df=pd.read_excel('../../data/3/pl_statics/'+str(year)+'.xlsx')
    lib_df['시점'] = year
    lib_df_list.append(lib_df)
    
lib_dfs = pd.concat(lib_df_list)
## 국공립 공공도서관 가져오기
lib_dfs = lib_dfs[(lib_dfs['설립주체']!='사립')].reset_index().drop(columns=['index','순번','연번'])

lib_dfs['지역'] = lib_dfs['지역'].apply(lambda x: sido_abbr_nm[x] if x in sido_abbr_nm else x)
lib_dfs['지역_시군구'] = lib_dfs.apply(lambda x:(city_sgg_change(x['지역'],x['지역_시군구'])), axis=1)

lib_dfs = lib_dfs.rename(columns={'지역':'시도','지역_시군구':'시군구'})

lib_dfs

lib_dfs_sgg = pd.merge(sgg_df, lib_dfs, on=['시도','시군구'],  how='inner')
lib_dfs_sgg

## 도서관 년식
lib_dfs_sgg['AGE'] = lib_dfs_sgg['시점'] - lib_dfs_sgg['개관년도'] + 1

## 소장 도서
lib_dfs_sgg['도서_총류'] = lib_dfs_sgg[['도서_국내_총류','도서_국외_총류']].sum(axis=1)
lib_dfs_sgg['도서_철학'] = lib_dfs_sgg[['도서_국내_철학','도서_국외_철학']].sum(axis=1)
lib_dfs_sgg['도서_종교'] = lib_dfs_sgg[['도서_국내_종교','도서_국외_종교']].sum(axis=1)
lib_dfs_sgg['도서_사회과학'] = lib_dfs_sgg[['도서_국내_사회과학','도서_국외_사회과학']].sum(axis=1)
lib_dfs_sgg['도서_순수과학'] = lib_dfs_sgg[['도서_국내_순수과학','도서_국외_순수과학']].sum(axis=1)
lib_dfs_sgg['도서_기술과학'] = lib_dfs_sgg[['도서_국내_기술과학','도서_국외_기술과학']].sum(axis=1)
lib_dfs_sgg['도서_예술'] = lib_dfs_sgg[['도서_국내_예술','도서_국외_예술']].sum(axis=1)
lib_dfs_sgg['도서_언어'] = lib_dfs_sgg[['도서_국내_언어','도서_국외_언어']].sum(axis=1)
lib_dfs_sgg['도서_문학'] = lib_dfs_sgg[['도서_국내_문학','도서_국외_문학']].sum(axis=1)
lib_dfs_sgg['도서_역사'] = lib_dfs_sgg[['도서_국내_역사','도서_국외_역사']].sum(axis=1)

lib_dfs_sgg['도서_국내'] = lib_dfs_sgg[['도서_국내_총류','도서_국내_철학','도서_국내_종교', '도서_국내_사회과학', '도서_국내_순수과학', '도서_국내_기술과학', '도서_국내_예술', '도서_국내_언어', '도서_국내_문학','도서_국내_역사']].sum(axis=1)
lib_dfs_sgg['도서_국외'] = lib_dfs_sgg[['도서_국외_총류','도서_국외_철학','도서_국외_종교', '도서_국외_사회과학', '도서_국외_순수과학', '도서_국외_기술과학', '도서_국외_예술', '도서_국외_언어', '도서_국외_문학','도서_국외_역사']].sum(axis=1)

lib_dfs_sgg['도서'] = lib_dfs_sgg[['도서_국내','도서_국외']].sum(axis=1)

## 대출권수 
lib_dfs_sgg['대출권수_총류'] = lib_dfs_sgg[['대출권수_어린이_총류','대출권수_청소년_총류', '대출권수_성인_총류']].sum(axis=1)
lib_dfs_sgg['대출권수_철학'] = lib_dfs_sgg[['대출권수_어린이_철학','대출권수_청소년_철학', '대출권수_성인_철학']].sum(axis=1)
lib_dfs_sgg['대출권수_종교'] = lib_dfs_sgg[['대출권수_어린이_종교','대출권수_청소년_종교', '대출권수_성인_종교']].sum(axis=1)
lib_dfs_sgg['대출권수_사회과학'] = lib_dfs_sgg[['대출권수_어린이_사회과학','대출권수_청소년_종교', '대출권수_성인_종교']].sum(axis=1)
lib_dfs_sgg['대출권수_순수과학'] = lib_dfs_sgg[['대출권수_어린이_순수과학','대출권수_청소년_순수과학', '대출권수_성인_순수과학']].sum(axis=1)
lib_dfs_sgg['대출권수_기술과학'] = lib_dfs_sgg[['대출권수_어린이_기술과학','대출권수_청소년_기술과학', '대출권수_성인_기술과학']].sum(axis=1)
lib_dfs_sgg['대출권수_예술'] = lib_dfs_sgg[['대출권수_어린이_예술','대출권수_청소년_예술', '대출권수_성인_예술']].sum(axis=1)
lib_dfs_sgg['대출권수_언어'] = lib_dfs_sgg[['대출권수_어린이_언어','대출권수_청소년_언어', '대출권수_성인_언어']].sum(axis=1)
lib_dfs_sgg['대출권수_문학'] = lib_dfs_sgg[['대출권수_어린이_문학','대출권수_청소년_문학', '대출권수_성인_문학']].sum(axis=1)
lib_dfs_sgg['대출권수_역사'] = lib_dfs_sgg[['대출권수_어린이_역사','대출권수_청소년_역사', '대출권수_성인_역사']].sum(axis=1)

lib_dfs_sgg['대출권수'] = lib_dfs_sgg[['대출권수_총류','대출권수_철학', '대출권수_종교', '대출권수_사회과학', '대출권수_순수과학', '대출권수_기술과학', '대출권수_예술', '대출권수_문학', '대출권수_역사']].sum(axis=1)

## 정규직 직원수
lib_dfs_sgg['직원_정규직'] = lib_dfs_sgg[['직원_정규직_사서남','직원_정규직_사서여','직원_정규직_행정남','직원_정규직_행정여','직원_정규직_전산남','직원_정규직_전산여', '직원_정규직_기타남', '직원_정규직_기타여']].sum(axis=1)

lib_dfs_sgg.head()

agg_dict = {
    '도서관명':'count',
    '시설_연면적' : 'mean', 
    '시설_연면적' : 'mean', 
    '개관일수': 'mean',
    'AGE': 'mean',
    '직원_정규직' :'sum',
    '도서관총예산' : 'sum',
    '도서_총류' :'sum',
    '도서_철학' :'sum', 
    '도서_종교' :'sum',
    '도서_사회과학' :'sum',
    '도서_순수과학' :'sum',
    '도서_기술과학' :'sum',
    '도서_예술' :'sum', 
    '도서_언어': 'sum',
    '도서_문학' :'sum',
    '도서_역사':'sum', 
    '도서_국내':'sum',
    '도서_국외':'sum',
    '도서':'sum',
    '대출권수_총류':'sum', 
    '대출권수_철학':'sum',
    '대출권수_종교':'sum',
    '대출권수_사회과학':'sum',
    '대출권수_순수과학':'sum',
    '대출권수_기술과학':'sum',
    '대출권수_예술':'sum',
    '대출권수_언어':'sum',
    '대출권수_문학':'sum',
    '대출권수_역사':'sum',
    '대출권수':'sum',   
}

lib_ele_dfs = lib_dfs_sgg[list(agg_dict.keys())]

lib_dfs_sgg = lib_dfs_sgg.groupby(['시군구코드','시도','시군구', '시점']).agg(agg_dict)
lib_dfs_sgg = lib_dfs_sgg.rename(columns={'도서관명':'도서관수'})
lib_dfs_sgg = lib_dfs_sgg.reset_index()
lib_dfs_sgg

lib_dfs_sgg.to_excel('../data/3/result/sgg_lib.xlsx',index=False)

lib_ele_dfs.head()

lib_ele_dfs.to_excel('../data/3/result/lib.xlsx',index=False)
