==> ipynb/4_data_merge.ipynb <==
import pandas as pd
import numpy as np

sdi_df = pd.read_excel('../data/1/result/sdi.xlsx', dtype={'시군구코드':str})
sdi_df

comm_control_var_df = pd.read_excel('../data/2/result/comm_control_var.xlsx', dtype={'시군구코드':str, '법정_시군구코드': str})
comm_control_var_df = comm_control_var_df[['시군구코드', '시점', '총인구수', '영유아_어린이_인구_비율', '청소년_인구_비율',
       '고령인구_비율', '여성인구비율', '자연인구증가율', '인위인구증가율', '십만명당_문화기반시설수', '15분이내_접근가능인구_비율_고등학교', '15분이내_접근가능인구_비율_공공의료시설',
        '15분이내_접근가능인구_비율_대규모점포', '15분이내_접근가능인구_비율_병·의원', '15분이내_접근가능인구_비율_전통시장', '15분이내_접근가능인구_비율_종합병원',
       '15분이내_접근가능인구_비율_중학교', '15분이내_접근가능인구_비율_초등학교',
       '재정자립도', '면적', '공원면적']]
comm_control_var_df

lib_df = pd.read_excel('../data/3/result/sgg_lib.xlsx', dtype={'시군구코드':str})

lib_df = lib_df[['시군구코드', '시점', '도서관수', '시설_연면적', '개관일수', 'AGE', '직원_정규직',
       '도서관총예산', '도서_총류', '도서_철학', '도서_종교', '도서_사회과학', '도서_순수과학', '도서_기술과학',
       '도서_예술', '도서_언어', '도서_문학', '도서_역사', '도서_국내', '도서_국외', '도서', '대출권수_총류',
       '대출권수_철학', '대출권수_종교', '대출권수_사회과학', '대출권수_순수과학', '대출권수_기술과학', '대출권수_예술',
       '대출권수_언어', '대출권수_문학', '대출권수_역사', '대출권수']]

lib_df

dataset = pd.merge(sdi_df, comm_control_var_df, on=['시군구코드','시점'], how='inner')
dataset = pd.merge(dataset, lib_df, on=['시군구코드','시점'], how='inner')

dataset

dataset[dataset['대출권수']==0]

dataset.to_excel('../data/4/result/dataset.xlsx', index=False)

## 시각화

import geopandas as gpd

sgg_gdf = gpd.gpd.GeoDataFrame.from_file('../data/0/shp/sgg.shp', encoding='UTF-8', crs="epsg:4326")
sgg_gdf = sgg_gdf.astype({'SGG_AD_CD':str,'SGG_LE_CD':str})
sgg_gdf

import matplotlib
from matplotlib.colors import TwoSlopeNorm, BoundaryNorm
import matplotlib.font_manager as fm
import matplotlib.pyplot as plt

fm.get_fontconfig_fonts()
font_location = 'C:\\Windows\\Fonts\malgun.ttf'
font_name = fm.FontProperties(fname=font_location).get_name()
matplotlib.rc('font', family=font_name)

import seaborn as sns

sns.set_style("whitegrid")
sns.set_context("poster")

plt.ioff()

for year in range(2014, 2019):

    year_data = dataset.loc[dataset['시점']==year].reset_index()
    year_data['1인당_대출권수'] = year_data['대출권수'] / year_data['총인구수'] 

    year_data_visu = gpd.GeoDataFrame(pd.merge(sgg_gdf, year_data, left_on ='SGG_AD_CD', right_on='시군구코드', how='inner'))

    # 10분위 수로 색을 칠하기
    bounds = np.percentile(year_data['1인당_대출권수'], np.linspace(0, 100, 11), interpolation='nearest')
    norm = BoundaryNorm(boundaries=bounds, ncolors=256)

    # create a normalized colorbar
    cmap = 'YlGn'
    cbar = plt.cm.ScalarMappable(norm=norm, cmap=cmap)

    # 읍면동 시각화
    ax = year_data_visu.plot(
        figsize=(50, 50), column='1인당_대출권수', cmap=cmap, norm=norm, legend=False)
    ax.set_axis_off()

    plt.savefig(f'../img/4/loan/{str(year)}.png', bbox_inches = 'tight', pad_inches = 0)
