==> ipynb/6_Analysis_Regression.ipynb <==
import statsmodels.api as sm
import statsmodels.formula.api as smf

import pandas as pd
import numpy as np

dataset = pd.read_excel('../data/4/result/dataset.xlsx')
dataset = dataset.set_index(['시군구', '시군구코드'])

dataset['도서관당_면적'] = dataset['면적'] / dataset['도서관수']
dataset['도서관당_도서'] = dataset['도서'] / dataset['도서관수']

dataset['일인당_대출권수'] = dataset['대출권수'] / dataset['총인구수']
# 시점을 더미 데이터로 변환 
dataset = pd.get_dummies(dataset, columns = ['시점'])
dataset = dataset.rename(columns={'15분이내_접근가능인구_비율_초등학교':'접근가능인구_비율_초등학교', '15분이내_접근가능인구_비율_공공의료시설':'접근가능인구_비율_공공의료시설'})

dataset.head(5)

dataset.dtypes

dataset.columns

regression_dataset = dataset[['시도', 'SDI', '총인구수', '영유아_어린이_인구_비율', '고령인구_비율', '여성인구비율', '자연인구증가율', '인위인구증가율', '십만명당_문화기반시설수', '접근가능인구_비율_공공의료시설', '재정자립도', '도서관당_면적', '도서','AGE', '시점_2014', '시점_2015', '시점_2016', '시점_2017', '시점_2018','공원면적', '재정자립도','직원_정규직', '일인당_대출권수']]
regression_dataset.head()

scale_var = []
categoty_var = ['시점_2014', '시점_2015','시점_2016', '시점_2017', '시점_2018']
just_var = ['SDI', '도서', '총인구수', '영유아_어린이_인구_비율', '고령인구_비율', '여성인구비율', '인위인구증가율', '직원_정규직', '접근가능인구_비율_공공의료시설']

cate_fomul = ' + '.join(["({})".format(name) for name in categoty_var])
cate_fomul

scale_fomul = ' + '.join(["scale({})".format(name) for name in scale_var])
scale_fomul

just_fomul = ' + '.join(["({})".format(name) for name in just_var])
just_fomul

indepent_fomul =  just_fomul+ ' + '  + cate_fomul 
indepent_fomul

# mixed effect model
md = smf.mixedlm("일인당_대출권수 ~ "+indepent_fomul, regression_dataset, groups=regression_dataset["시도"])

mdf = md.fit()

print(mdf.summary())
